+++ modules/anime/src/main/kotlin/com/thekeeperofpie/artistalleydatabase/anime/AnimeSettings.kt
@@ -2,6 +2,7 @@ package com.thekeeperofpie.artistalleydatabase.anime
 
 import com.anilist.type.MediaType
 import com.thekeeperofpie.artistalleydatabase.anilist.AniListLanguageOption
+import com.thekeeperofpie.artistalleydatabase.anilist.VoiceActorLanguageOption
 import com.thekeeperofpie.artistalleydatabase.anime.media.filter.FilterData
 import com.thekeeperofpie.artistalleydatabase.anime.media.ui.MediaViewOption
 import com.thekeeperofpie.artistalleydatabase.anime.news.AnimeNewsNetworkCategory
@@ -38,6 +39,8 @@ interface AnimeSettings {
     val languageOptionMedia: MutableStateFlow<AniListLanguageOption>
     val languageOptionCharacters: MutableStateFlow<AniListLanguageOption>
     val languageOptionStaff: MutableStateFlow<AniListLanguageOption>
+    val languageOptionVoiceActor: MutableStateFlow<VoiceActorLanguageOption>
+    val showFallbackVoiceActor: MutableStateFlow<Boolean>
 
     val lastCrash: MutableStateFlow<String>
     val lastCrashShown: MutableStateFlow<Boolean>