+++ changelog.md
@@ -4,6 +4,7 @@
 - Add full current media screen
 - Add language setting to search filters
 - Add scroll bars to user list
+- Add voice actor language option
 
 ### 0.42 (7)
 - Add unread notification badge